clean up runQueueConsumer configurations in queue tests
